Make R linter `lintr` stateless and efficient
The call to R should respect `inferior-R-args` variable. If not, then at least please do not call the R slave session with `--restore`, which would try to load `./RData`, and depending on what `PWD` is, that could try to load gigabytes of workspace image into the lint, including variables and functions. That introduces unacceptable lag (every keystore requires over 10s of wait time whenever the linter is invoked then killed off, loading gigabytes of image). What's worse is that a linter should start stateless and thus be predictable; loading arbitrary images from a "current directory or from a default fallback location" is a very bad behavior for a linter.