Fix message truncation for python-mypy due to pretty formatting
If the the user has "pretty = true" in their mypy config it can result in
truncation of the messages reported. We invoke mypy with --no-pretty to override
this.

Minimum version is bumped to 0.730 (2019-09-26) because this is the version in
which --pretty and --no-pretty were added:
https://mypy-lang.blogspot.com/2019/09/mypy-730-released.html

Fixes GH-1983 and closes GH-1983