bug: fix arrows directions
When setting fringe or border to 'right', the provided marker kept
pointing right.

This aims to change the default behaviour of
`flycheck-redefine-standard-error-levels` to make into account the
selected `flycheck-indication-mode`.

So if the left-margin or left-fringe is set, everything stays the
same, buf if they're set to the right-margin or rigt-fringe an arrow
pointing to the left will be used.

Diff may show other little fixes, this was done by `make format`, I
thought there's nothing bad about it and kept it on this commit.