docstrings should not exceed 80 characters in width
The docstring generated for flycheck-stylelintrc includes each of its
checkers, ending up 87 characters long.  I restricted its definition to print
only the first three (of 4) checkers, which is enough to fix it, though we
might want to consider putting this limit in the macro that generates the
docstring.

Separately, the docstring for flycheck-cuda-extended-lambda included an
overly-long line, which I wrapped.