Add a new checker 'r' for R
This is a rather rudimentary checker using only the builtin 'parse'
function of R. It can only report syntax/parsing errors, but it does not
depend on any third-party libraries.

R actually has some builtin utilities for linting (for example,
'codetools'), but as far as I know those require evaluating the source
code, so only 'parse' is used.